refactor(board): extract element helpers
